+++ src/styles/ommon.tsx
